refactor: extract deployment spec crate
