+++ Cargo.toml
@@ -3,6 +3,7 @@ members = [
     ".",
     "crates/rise-backend-auth",
     "crates/rise-backend-core",
+    "crates/rise-deployment-spec",
     "crates/rise-resource-api",
     "crates/rise-resource-store",
     "crates/rise-runtime-sync",
@@ -28,6 +29,7 @@ sha2 = "0.11"
 sqlx = { version = "0.8", features = ["runtime-tokio-native-tls", "postgres", "uuid", "chrono", "migrate", "json"] }
 thiserror = "2.0"
 tokio = { version = "1.48.0", features = ["full"] }
+rise-deployment-spec = { path = "crates/rise-deployment-spec" }
 tracing = "0.1.43"
 uuid = { version = "1.18.1", features = ["v4", "serde"] }
 
@@ -100,6 +102,7 @@ backend = [
     "dep:schemars",
     "dep:serde_yaml",
     "dep:tokio-tungstenite",
+    "rise-deployment-spec/schema",
     # Kubernetes (former "k8s" feature)
     "dep:kube",
     "dep:k8s-openapi",
@@ -131,6 +134,7 @@ rise-runtime-sync = { path = "crates/rise-runtime-sync", optional = true }
 
 # Common (always enabled - shared across features)
 anyhow = "1.0.100"
+rise-deployment-spec = { workspace = true }
 async-stream = "0.3"
 base64 = { workspace = true }
 bytes = "1.9"
