+++ crates/rise-deployment-spec/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "rise-deployment-spec"
+version = "0.1.0"
+edition = "2021"
+license = "MIT OR Apache-2.0"
+
+[features]
+default = []
+schema = ["dep:schemars"]
+
+[dependencies]
+schemars = { workspace = true, optional = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+
+[dev-dependencies]
+toml = "1"
