+++ crates/rise-deployment-spec/src/lib.rs
@@ -1,0 +1,8 @@
+//! Shared deployment specification and project configuration model for Rise.
+
+pub mod project_config;
+pub mod request_spec;
+pub mod side_data;
+pub mod validation;
+
+pub use project_config::*;
